Use datetime.time.max for end date range filters
When passing a date to a DateTimeType range filter, we used
to set the time to midnight for both the start and the end range
filters. So the following search parameters

`start__gte='2019-12-25', start__lte='2019-12-25'`

would be converted in the sdk to

`'start.gte': '2016-12-25T00:00:00.000000', 'start.lte': '2016-12-25T00:00:00.000000'`

This would exclude an event with the following start value

`'start': '2016-12-25T05:00:00.000000'

This commits sets the end filters (lt, lte) to the datetime max for the day,
i.e., `23:59:59.999999`. So the parameters mentioned above would now get converted to

`'start.gte': '2016-12-25T00:00:00.000000', 'start.lte': '2016-12-25T23:59:59.999999'`

and would include the event mentioned above.